Moved ImageUploaded logic from CameraCapture into its own file
